Revert "correct type of vertex weight array"
This reverts commit 1014dafc06a7ba7218ce408582882f5949c991e1.

+++ src/graph/transform/metis_partition_hetero.cc
@@ -40,7 +40,7 @@ IdArray KaHIPPartition(
       << "The vertex weight array doesn't have right type";
   CHECK(vwgt_len % g->NumVertices(0) == 0)
       << "The vertex weight array doesn't have right number of elements";
-  int64_t *vwgt = NULL;
+  int *vwgt = NULL;
   if (vwgt_len > 0) {
     vwgt = static_cast<int *>(vwgt_arr->data);
   }
